src/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cmake==3.30.5
contourpy==1.3.0
cycler==0.12.1
dataclasses==0.6
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.1
idna==3.10
imutils==0.5.4
intel-cmplr-lib-ur==2025.0.0
intel-openmp==2025.0.0
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mkl==2025.0.0
mkl-include==2025.0.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
slack_sdk==3.33.3
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.1
tbb==2022.0.0
tcmlib==1.2.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1
tqdm==4.66.6
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
umf==0.9.0
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.2
